Revise for review comments 1. create separate file for NextHopKey and NextHopGroupKey 2. remove test_AclRuleRedirectToNexthop
