+++ orchagent/nexthopkey.h
@@ -1,0 +1,67 @@
+#ifndef SWSS_NEXTHOPKEY_H
+#define SWSS_NEXTHOPKEY_H
+
+#include "ipaddress.h"
+#include "tokenize.h"
+
+#define VRF_PREFIX "Vrf"
+extern IntfsOrch *gIntfsOrch;
+
+struct NextHopKey
+{
+    IpAddress           ip_address;     // neighbor IP address
+    string              alias;          // incoming interface alias
+
+    NextHopKey() = default;
+    NextHopKey(const std::string &ipstr, const std::string &alias) : ip_address(ipstr), alias(alias) {}
+    NextHopKey(const IpAddress &ip, const std::string &alias) : ip_address(ip), alias(alias) {}
+    NextHopKey(const std::string &str)
+    {
+        if (str.find(',') != string::npos)
+        {
+            std::string err = "Error converting " + str + " to NextHop";
+            throw std::invalid_argument(err);
+        }
+        auto keys = tokenize(str, '|');
+        if (keys.size() == 1)
+        {
+            ip_address = keys[0];
+            alias = gIntfsOrch->getRouterIntfsAlias(ip_address);
+        }
+        else if (keys.size() == 2)
+        {
+            ip_address = keys[0];
+            alias = keys[1];
+            if (!alias.compare(0, strlen(VRF_PREFIX), VRF_PREFIX))
+            {
+                alias = gIntfsOrch->getRouterIntfsAlias(ip_address, alias);
+            }
+        }
+        else
+        {
+            std::string err = "Error converting " + str + " to NextHop";
+            throw std::invalid_argument(err);
+        }
+    }
+    const std::string to_string() const
+    {
+        return ip_address.to_string() + "|" + alias;
+    }
+
+    bool operator<(const NextHopKey &o) const
+    {
+        return tie(ip_address, alias) < tie(o.ip_address, o.alias);
+    }
+
+    bool operator==(const NextHopKey &o) const
+    {
+        return (ip_address == o.ip_address) && (alias == o.alias);
+    }
+
+    bool operator!=(const NextHopKey &o) const
+    {
+        return !(*this == o);
+    }
+};
+
+#endif /* SWSS_NEXTHOPKEY_H */
